Update script.js
going to try to create a frankenstein of all of the best bits of the last 3 attempts and try to pass the tests, so unbeleivably annoying, it SHOULD work, but it doesn't, even the example project code can't pass it's own tests, is the tester working properly? am I updating the div correctly? I've had many problems with the tests before, but I do find it hilarious how the code they use as an example, doesn't work. what hope does my frankenstein code have, if the 'correct' code doesn't even work. It's not even that I can't fix the problem, I don't even know what the problem is that needs to be fixed because JS can't be debugged in VScode and I reallllllllllllly don't feell like learning node yet for a simple AF JS project