Update script.js
The Test:
When you click on the #check-btn element without entering a value into the #user-input element, an alert should appear with the text Please provide a phone number.

HOT GARBAGE TESTING, even if i explicitly tell it wtf to do, it will still fail. leaving me dejected, angry, and confused.

That's just -1- test, and I can't even get it to pass that, what is going on